@sam-m888 2017-11-12
Worked for my 13000 person family tree.