@prculley 2017-11-17
@sam-m888 When I redid it, I compared the counts to the original version as running under Gramps4.2.6.  They are the same.  I would update the wiki to match the results.