@sam-m888 2017-11-17
@prculley Ok, I've made the wiki description simpler and removed that line.